Added documentation
Necessary to explain what this gem is and why it's important.

This is a direct port of the original Wholable gem -- which is deprecated and will soon be removed -- in order to fix the name (was missing an "e") and be more consistent with gems, like link:https://alchemists.io/projects/pipeable[Pipeable], to reduce confusion and improve readability.