Added members
Necessary to provide compatibility with `Data` and `Struct` object APIs by allowing you to retrieve all members of the object as originally defined. This provides several benefits:

* Simplifies the implementation, by refactoring, to use the new `#members` instance method instead of supplying the member keys to each `#define_` function.
* Ensures inheritance is accounted for by ensuring the subclass accumulates all members from the superclass in order to dynamically grow the members as new subclasses add new attributes.

Reek was updated to remove the feature envy exclusion since it's no longer needed and ignore the manual dispatch violation because the only way to know if the `descendant` is a wholeable object is to check if it responds to `.members`. There are other ways to solve this but, in this particular case, it's best to make this calculation once and then supply the members array to both the `.members` and `#members` methods so the subclass doesn't have to make this calculation over and over again which improves performance.

Milestone: minor